+++ protobuf/lib/src/protobuf/coded_buffer_writer.dart
@@ -22,8 +22,8 @@ part of '../../protobuf.dart';
 class CodedBufferWriter {
   /// Array of splices representing the data written into the writer.
   /// Each element might be one of:
-  ///   * a TypedData object - represents a sequence of bytes that need to be
-  ///                          emitted into the result as-is;
+  ///   * a [Uint8List] object - represents a sequence of bytes that need to be
+  ///                            emitted into the result as-is;
   ///   * a positive integer - a number of bytes to copy from [_outputChunks]
   ///                          into resulting buffer;
   ///   * a non-positive integer - a positive number that needs to be emitted
@@ -141,21 +141,21 @@ class CodedBufferWriter {
           }
           buffer[outPos++] = v;
         } else {
-          // action is an amount of bytes to copy from _outputChunks into the
-          // buffer.
+          // `action` is an amount of bytes to copy from `_outputChunks` into
+          // the buffer.
           var bytesToCopy = action;
           while (bytesToCopy > 0) {
             final Uint8List chunk = _outputChunks[chunkIndex];
             final int bytesInChunk = _outputChunks[chunkIndex + 1];
 
-            // Copy at most bytesToCopy bytes from the current chunk.
+            // Copy at most `bytesToCopy` bytes from the current chunk.
             final leftInChunk = bytesInChunk - chunkPos;
             final bytesToCopyFromChunk =
                 leftInChunk > bytesToCopy ? bytesToCopy : leftInChunk;
-            final endPos = chunkPos + bytesToCopyFromChunk;
-            while (chunkPos < endPos) {
-              buffer[outPos++] = chunk[chunkPos++];
-            }
+            buffer.setRange(
+                outPos, outPos + bytesToCopyFromChunk, chunk, chunkPos);
+            chunkPos += bytesToCopyFromChunk;
+            outPos += bytesToCopyFromChunk;
             bytesToCopy -= bytesToCopyFromChunk;
 
             // Move to the next chunk if the current one is exhausted.
@@ -166,9 +166,12 @@ class CodedBufferWriter {
           }
         }
       } else {
-        // action is a TypedData containing bytes to emit into the output
+        // action is a `Uint8List` containing bytes to emit into the output
         // buffer.
-        outPos = _copyInto(buffer, outPos, action);
+        final Uint8List value = action;
+        final end = outPos + value.length;
+        buffer.setRange(outPos, end, value);
+        outPos = end;
       }
     }
 
@@ -210,7 +213,7 @@ class CodedBufferWriter {
   /// Record number of bytes written into output chunks since last splice.
   ///
   /// This is used before reserving space for an unknown varint splice or
-  /// adding a TypedData array splice.
+  /// adding a [Uint8List] array splice.
   void _commitSplice() {
     final pos = _bytesInChunk + _outputChunksBytes;
     final bytes = pos - _lastSplicePos;
@@ -220,12 +223,14 @@ class CodedBufferWriter {
     _lastSplicePos = pos;
   }
 
-  /// Add TypedData splice - these bytes would be directly copied into the
-  /// output buffer by [writeTo].
-  void writeRawBytes(TypedData value) {
+  /// Add a [Uint8List] splice, without copying. These bytes will be directly
+  /// copied into the output buffer by [writeTo].
+  void writeRawBytes(Uint8List value) {
+    final length = value.lengthInBytes;
+    if (length == 0) return;
     _commitSplice();
     _splices.add(value);
-    _bytesTotal += value.lengthInBytes;
+    _bytesTotal += length;
   }
 
   /// Start writing a length-delimited data.
@@ -337,11 +342,22 @@ class CodedBufferWriter {
         _writeVarint32(value ? 1 : 0);
         break;
       case PbFieldType._BYTES_BIT:
-        _writeBytesNoTag(
-            value is Uint8List ? value : Uint8List.fromList(value));
+        final List<int> bytes = value;
+        if (bytes is Uint8List) {
+          _writeBytesNoTag(bytes);
+        } else if (bytes.isEmpty) {
+          _writeEmptyBytes();
+        } else {
+          _writeBytesNoTag(Uint8List.fromList(bytes));
+        }
         break;
       case PbFieldType._STRING_BIT:
-        _writeBytesNoTag(_utf8.encoder.convert(value));
+        final String string = value;
+        if (string.isEmpty) {
+          _writeEmptyBytes();
+        } else {
+          _writeBytesNoTag(const Utf8Encoder().convert(string));
+        }
         break;
       case PbFieldType._DOUBLE_BIT:
         _writeDouble(value);
@@ -354,9 +370,19 @@ class CodedBufferWriter {
         _writeVarint32(enum_.value & 0xffffffff);
         break;
       case PbFieldType._GROUP_BIT:
-        // value is UnknownFieldSet or GeneratedMessage
-        // ignore: avoid_dynamic_calls
-        value.writeToCodedBufferWriter(this);
+        // `value` is `UnknownFieldSet` or `GeneratedMessage`. Test for
+        // `UnknownFieldSet` as it doesn't have subtypes, so the type test will
+        // be fast.
+        if (value is UnknownFieldSet) {
+          // Give the variable a type to not rely on type promotion to
+          // eliminate the dynamic call below.
+          // ignore: omit_local_variable_types
+          final UnknownFieldSet unknownFieldSet = value;
+          unknownFieldSet.writeToCodedBufferWriter(this);
+        } else {
+          final GeneratedMessage message = value;
+          message.writeToCodedBufferWriter(this);
+        }
         break;
       case PbFieldType._INT32_BIT:
         _writeVarint64(Int64(value));
@@ -402,6 +428,10 @@ class CodedBufferWriter {
     writeRawBytes(value);
   }
 
+  void _writeEmptyBytes() {
+    writeInt32NoTag(0);
+  }
+
   void _writeTag(int fieldNumber, int wireFormat) {
     writeInt32NoTag(makeTag(fieldNumber, wireFormat));
   }
@@ -417,27 +447,6 @@ class CodedBufferWriter {
 
   void writeInt32NoTag(int value) {
     _writeVarint32(value & 0xFFFFFFFF);
-  }
-
-  /// Copy bytes from the given typed data array into the output buffer.
-  ///
-  /// Has a specialization for Uint8List for performance.
-  int _copyInto(Uint8List buffer, int pos, TypedData value) {
-    if (value is Uint8List) {
-      final len = value.length;
-      for (var j = 0; j < len; j++) {
-        buffer[pos++] = value[j];
-      }
-      return pos;
-    } else {
-      final len = value.lengthInBytes;
-      final u8 = Uint8List.view(
-          value.buffer, value.offsetInBytes, value.lengthInBytes);
-      for (var j = 0; j < len; j++) {
-        buffer[pos++] = u8[j];
-      }
-      return pos;
-    }
   }
 
   /// This function maps a power-of-2 value (2^0 .. 2^31) to a unique value