+++ protoc_plugin/lib/src/options.dart
@@ -50,8 +50,14 @@ bool genericOptionsParser(CodeGeneratorRequest request,
 class GenerationOptions {
   final bool useGrpc;
   final bool generateMetadata;
+  final bool disableConstructorArgs;
+  final bool useNullable;
 
-  GenerationOptions({this.useGrpc = false, this.generateMetadata = false});
+  GenerationOptions(
+      {this.useGrpc = false,
+      this.generateMetadata = false,
+      this.disableConstructorArgs = false,
+      this.useNullable = false});
 }
 
 /// A parser for a name-value pair option. Options parsed in
@@ -84,10 +90,36 @@ class GenerateMetadataParser implements SingleOptionParser {
   @override
   void parse(String name, String? value, OnError onError) {
     if (value != null) {
-      onError('Invalid metadata option. No value expected.');
+      onError('Invalid generate_kythe_info option. No value expected.');
       return;
     }
     generateKytheInfo = true;
+  }
+}
+
+class DisableConstructorArgsParser implements SingleOptionParser {
+  bool value = false;
+
+  @override
+  void parse(String name, String? value, OnError onError) {
+    if (value != null) {
+      onError('Invalid disable_constructor_args option. No value expected.');
+      return;
+    }
+    this.value = true;
+  }
+}
+
+class NullableOptionParser implements SingleOptionParser {
+  bool nullableEnabled = false;
+
+  @override
+  void parse(String name, String? value, OnError onError) {
+    if (value != null) {
+      onError('Invalid nullable option. No value expected.');
+      return;
+    }
+    nullableEnabled = true;
   }
 }
 
@@ -102,13 +134,21 @@ GenerationOptions? parseGenerationOptions(
 
   final grpcOptionParser = GrpcOptionParser();
   newParsers['grpc'] = grpcOptionParser;
+
   final generateMetadataParser = GenerateMetadataParser();
   newParsers['generate_kythe_info'] = generateMetadataParser;
+  final nullableOptionParser = NullableOptionParser();
+  newParsers['nullable'] = nullableOptionParser;
+
+  final disableConstructorArgsParser = DisableConstructorArgsParser();
+  newParsers['disable_constructor_args'] = disableConstructorArgsParser;
 
   if (genericOptionsParser(request, response, newParsers)) {
     return GenerationOptions(
         useGrpc: grpcOptionParser.grpcEnabled,
-        generateMetadata: generateMetadataParser.generateKytheInfo);
+        generateMetadata: generateMetadataParser.generateKytheInfo,
+        disableConstructorArgs: disableConstructorArgsParser.value,
+        useNullable: nullableOptionParser.nullableEnabled);
   }
   return null;
 }