+++ protoc_plugin/lib/src/protobuf_field.dart
@@ -65,6 +65,12 @@ class ProtobufField {
   bool get isRepeated =>
       descriptor.label == FieldDescriptorProto_Label.LABEL_REPEATED;
 
+  bool get isOptional {
+    if (isRepeated) return false;
+    if (isRequired || !descriptor.proto3Optional) return false;
+    return true;
+  }
+
   /// Whether a numeric field is repeated and must be encoded with packed
   /// encoding.
   ///
@@ -146,7 +152,7 @@ class ProtobufField {
       final d = baseType.generator as MessageGenerator;
       final keyType = d._fieldList[0].baseType.getDartType(parent.fileGen!);
       final valueType = d._fieldList[1].baseType.getDartType(parent.fileGen!);
-      return '$coreImportPrefix.Map<$keyType, $valueType>';
+      return '$protobufImportPrefix.PbMap<$keyType, $valueType>';
     }
     if (isRepeated) return baseType.getRepeatedDartType(parent.fileGen!);
     return baseType.getDartType(parent.fileGen!);